rewrite plugin manager
split concept of plugin and command
