--- sigma/__init__.py
@@ -1,13 +1,0 @@
-from .sigma import Sigma
-from . import utils
-from . import database
-from . import plugin
-from . import plugin_manager
-
-__all__ = [
-    'Sigma',
-    'utils',
-    'database',
-    'plugin',
-    'plugin_manager'
-]
