+++ sigma/core/formatting.py
@@ -1,0 +1,22 @@
+def italics(text):
+    return '*{:s}*'.format(text)
+
+
+def bold(text):
+    return '**{:s}**'.format(text)
+
+
+def strikeout(text):
+    return '~~{:s}~~'.format(text)
+
+
+def underline(text):
+    return '__{:s}__'.format(text)
+
+
+def code(text):
+    return '`{:s}`'.format(text)
+
+
+def codeblock(text):
+    return '```{:s}```'.format(text)
