--- sigma/plugins/nihongo/__init__.py
@@ -1,12 +1,0 @@
-from .wanikani import WK
-from .wanikani import WKKey
-from .jisho import Jisho
-
-
-__all__ = [
-    'WK',
-    'WKKey',
-    'Jisho'
-]
-
-pluglist = __all__
