+++ sigma/plugins/nihongo/wanikani/logger.py
@@ -1,0 +1,3 @@
+from sigma.core.logger import create_logger
+
+log = create_logger('wanikani')
