+++ sigma/plugins/nihongo/wanikani/plugin.yml
@@ -1,0 +1,31 @@
+name: WaniKani
+
+access:
+  database: true
+
+categories:
+  - nihongo
+
+enabled: true
+
+commands:
+  - name:    wanikani
+    global:  true
+    enabled: true
+    sfw:     true
+    usage:   "{pfx:s}{cmd:s} breadstickninja"
+    description:
+      Shows the WaniKani statistics of the person who used the command or a
+      mentioned person. Also searches WaniKani for a given username.
+      For advanced stats, a user must save his WaniKani API Key with the
+      %pfx%wksave command. Use help on it to get more info.
+
+  - name:    wksave
+    global:  true
+    enabled: true
+    sfw:     true
+    usage:   "{pfx:s}{cmd:s} key 12345678901234567890123456789012"
+    description:
+      Stores a given API Key or Username for you.
+      If you want detailed stats when calling the %pfx%wanikani command
+      you need to save your WK API Key like in the example bellow!
